+++ .github/scripts/release/windows.ps1
@@ -1,6 +1,21 @@
 ## Copyright 2020 Intel Corporation
 ## SPDX-License-Identifier: Apache-2.0
 
+Param(
+  [string] $G = 'Visual Studio 17 2022'
+)
+
+# abort on any error
+$ErrorActionPreference = 'Stop'
+function Assert-Success { if ($LASTEXITCODE) { exit $LASTEXITCODE } }
+
+# Copy-Item silently ignores wildcards matching nothing
+function Copy-Required($Path, $Destination) {
+  $files = @(Get-ChildItem $Path)
+  if (!$files) { throw "nothing matches $Path" }
+  Copy-Item $files $Destination
+}
+
 #### Set variables for script ####
 
 $ROOT_DIR = pwd
@@ -12,22 +27,32 @@ $OPENVKL_PKG_BASE = "openvkl-$OPENVKL_RELEASE_PACKAGE_VERSION.x86_64.windows"
 $OPENVKL_BUILD_DIR = "$ROOT_DIR/build_release"
 $OPENVKL_INSTALL_DIR = "$ROOT_DIR/install_release/$OPENVKL_PKG_BASE"
 
-## Build dependencies ##
+$THREADS = $env:NUMBER_OF_PROCESSORS
+
+#### Cleanup any existing directories ####
+
+rm -Recurse -Force -ErrorAction SilentlyContinue `
+  $DEP_INSTALL_DIR, $DEP_BUILD_DIR, $OPENVKL_BUILD_DIR, $OPENVKL_INSTALL_DIR
+
+#### Build dependencies ####
 
 mkdir $DEP_BUILD_DIR
 cd $DEP_BUILD_DIR
 
 cmake --version
 
-cmake -L `
-  -G $args[0] `
+cmake `
+  $args `
+  -G $G `
   -D BUILD_DEPENDENCIES_ONLY=ON `
   -D BUILD_OPENVKL_BENCHMARKS=ON `
   -D CMAKE_INSTALL_PREFIX=$DEP_INSTALL_DIR `
   -D CMAKE_INSTALL_LIBDIR=lib `
   ../superbuild
+Assert-Success
 
-cmake --build . --config Release --target ALL_BUILD -- /m /nologo
+cmake --build . --config Release --parallel $THREADS
+Assert-Success
 
 cd $ROOT_DIR
 
@@ -36,15 +61,10 @@ cd $ROOT_DIR
 mkdir $OPENVKL_BUILD_DIR
 cd $OPENVKL_BUILD_DIR
 
-# Setup environment variables for dependencies
-$env:rkcommon_DIR = $DEP_INSTALL_DIR
-$env:embree_DIR = $DEP_INSTALL_DIR
-$env:glfw3_DIR = $DEP_INSTALL_DIR
-
 # set release settings
 cmake -L `
-  -G $args[0] `
-  -D CMAKE_PREFIX_PATH="$DEP_INSTALL_DIR\lib\cmake" `
+  -G $G `
+  -D CMAKE_PREFIX_PATH="$DEP_INSTALL_DIR" `
   -D CMAKE_INSTALL_PREFIX="$OPENVKL_INSTALL_DIR" `
   -D CMAKE_INSTALL_INCLUDEDIR=include `
   -D CMAKE_INSTALL_LIBDIR=lib `
@@ -53,18 +73,29 @@ cmake -L `
   -D RKCOMMON_TBB_ROOT=$DEP_INSTALL_DIR `
   -D ISPC_EXECUTABLE=$DEP_INSTALL_DIR/bin/ispc.exe `
   -D BUILD_BENCHMARKS=ON `
+  -D OpenVDB_ROOT=$DEP_INSTALL_DIR `
+  -D CMAKE_NO_SYSTEM_FROM_IMPORTED=ON `
   ..
+Assert-Success
 
-# build
-cmake --build . --config Release --target ALL_BUILD -- /m /nologo
-
-# install
-cmake --build . --config Release --target install -- /m /nologo
+# build and install
+cmake --build . --config Release --parallel $THREADS --target install
+Assert-Success
 
 # copy dependent libs into the install
 $INSTALL_BIN_DIR = "$OPENVKL_INSTALL_DIR/bin"
+$INSTALL_LIB_DIR = "$OPENVKL_INSTALL_DIR/lib"
 
-cp $DEP_INSTALL_DIR/bin/*.dll $INSTALL_BIN_DIR
+Copy-Required $DEP_INSTALL_DIR/bin/*.dll $INSTALL_BIN_DIR
+
+# openvklConfig.cmake globs for the dependencies' import libs
+Copy-Required $DEP_INSTALL_DIR/lib/rkcommon*.lib $INSTALL_LIB_DIR
+Copy-Required $DEP_INSTALL_DIR/lib/embree*.lib $INSTALL_LIB_DIR
+Copy-Required $DEP_INSTALL_DIR/lib/tbb*.lib $INSTALL_LIB_DIR
+
+# the debug variants of TBB are not needed
+rm $INSTALL_BIN_DIR/*_debug.dll
+rm $INSTALL_LIB_DIR/*_debug.lib
 
 # sign
 ;& $env:SIGN_FILE_WINDOWS -q -vv (Get-ChildItem $INSTALL_BIN_DIR\* | Select-Object -Expand FullName)