+++ .github/scripts/release/windows_sycl.ps1
@@ -1,6 +1,17 @@
 ## Copyright 2023 Intel Corporation
 ## SPDX-License-Identifier: Apache-2.0
 
+# abort on any error
+$ErrorActionPreference = 'Stop'
+function Assert-Success { if ($LASTEXITCODE) { exit $LASTEXITCODE } }
+
+# Copy-Item silently ignores wildcards matching nothing
+function Copy-Required($Path, $Destination) {
+  $files = @(Get-ChildItem $Path)
+  if (!$files) { throw "nothing matches $Path" }
+  Copy-Item $files $Destination
+}
+
 #### Set variables for script ####
 
 $ROOT_DIR = pwd
@@ -12,24 +23,33 @@ $OPENVKL_PKG_BASE = "openvkl-$OPENVKL_RELEASE_PACKAGE_VERSION.sycl.x86_64.window
 $OPENVKL_BUILD_DIR = "$ROOT_DIR/build_release"
 $OPENVKL_INSTALL_DIR = "$ROOT_DIR/install_release/$OPENVKL_PKG_BASE"
 
-## Build dependencies ##
+$THREADS = $env:NUMBER_OF_PROCESSORS
+
+#### Cleanup any existing directories ####
+
+rm -Recurse -Force -ErrorAction SilentlyContinue `
+  $DEP_INSTALL_DIR, $DEP_BUILD_DIR, $OPENVKL_BUILD_DIR, $OPENVKL_INSTALL_DIR
+
+#### Build dependencies ####
 
 mkdir $DEP_BUILD_DIR
 cd $DEP_BUILD_DIR
 
 cmake --version
 
-cmake -L `
+cmake `
+  $args `
   -G Ninja `
   -D BUILD_DEPENDENCIES_ONLY=ON `
   -D BUILD_OPENVKL_BENCHMARKS=ON `
   -D CMAKE_INSTALL_PREFIX=$DEP_INSTALL_DIR `
   -D CMAKE_INSTALL_LIBDIR=lib `
   -D CMAKE_CXX_COMPILER=clang-cl -D CMAKE_C_COMPILER=clang-cl `
-  -D BUILD_OPENVDB=OFF `
   ../superbuild
+Assert-Success
 
-cmake --build . --config Release --verbose
+cmake --build . --config Release --parallel $THREADS
+Assert-Success
 
 cd $ROOT_DIR
 
@@ -38,15 +58,10 @@ cd $ROOT_DIR
 mkdir $OPENVKL_BUILD_DIR
 cd $OPENVKL_BUILD_DIR
 
-# Setup environment variables for dependencies
-$env:rkcommon_DIR = $DEP_INSTALL_DIR
-$env:embree_DIR = $DEP_INSTALL_DIR
-$env:glfw3_DIR = $DEP_INSTALL_DIR
-
 # set release settings
 cmake -L `
   -G Ninja `
-  -D CMAKE_PREFIX_PATH="$DEP_INSTALL_DIR\lib\cmake" `
+  -D CMAKE_PREFIX_PATH="$DEP_INSTALL_DIR" `
   -D CMAKE_INSTALL_PREFIX="$OPENVKL_INSTALL_DIR" `
   -D CMAKE_INSTALL_INCLUDEDIR=include `
   -D CMAKE_INSTALL_LIBDIR=lib `
@@ -58,33 +73,39 @@ cmake -L `
   -D BUILD_BENCHMARKS=ON `
   -D CMAKE_CXX_COMPILER=clang++ -D CMAKE_C_COMPILER=clang `
   -D OPENVKL_ENABLE_DEVICE_GPU=ON `
+  -D OpenVDB_ROOT=$DEP_INSTALL_DIR `
+  -D CMAKE_NO_SYSTEM_FROM_IMPORTED=ON `
   ..
+Assert-Success
 
-# build
-cmake --build . --config Release --verbose
-
-# install
-cmake --build . --config Release --target install
+# build and install
+cmake --build . --config Release --parallel $THREADS --target install
+Assert-Success
 
 # copy dependent libs into the install
 $INSTALL_BIN_DIR = "$OPENVKL_INSTALL_DIR/bin"
 $INSTALL_LIB_DIR = "$OPENVKL_INSTALL_DIR/lib"
 
-cp $DEP_INSTALL_DIR/bin/*.dll $INSTALL_BIN_DIR
+Copy-Required $DEP_INSTALL_DIR/bin/*.dll $INSTALL_BIN_DIR
 
-cp $DEP_INSTALL_DIR/lib/rkcommon*.lib $INSTALL_LIB_DIR
-cp $DEP_INSTALL_DIR/lib/embree*.lib $INSTALL_LIB_DIR
-cp $DEP_INSTALL_DIR/lib/tbb*.lib $INSTALL_LIB_DIR
+# openvklConfig.cmake globs for the dependencies' import libs
+Copy-Required $DEP_INSTALL_DIR/lib/rkcommon*.lib $INSTALL_LIB_DIR
+Copy-Required $DEP_INSTALL_DIR/lib/embree*.lib $INSTALL_LIB_DIR
+Copy-Required $DEP_INSTALL_DIR/lib/tbb*.lib $INSTALL_LIB_DIR
+
+# the debug variants of TBB are not needed
+rm $INSTALL_BIN_DIR/*_debug.dll
+rm $INSTALL_LIB_DIR/*_debug.lib
 
 # copy SYCL runtime dependencies
 $SYCL_BIN_FILE = (get-command clang++).Path
 $SYCL_BIN_DIR = Split-Path -Parent "$SYCL_BIN_FILE"
 
-cp ${SYCL_BIN_DIR}/sycl?.dll $INSTALL_BIN_DIR
-cp ${SYCL_BIN_DIR}/ur_loader.dll $INSTALL_BIN_DIR
-cp ${SYCL_BIN_DIR}/ur_adapter_level_zero.dll $INSTALL_BIN_DIR
-cp ${SYCL_BIN_DIR}/ur_adapter_level_zero_v?.dll $INSTALL_BIN_DIR
-cp ${SYCL_BIN_DIR}/ur_win_proxy_loader.dll $INSTALL_BIN_DIR
+Copy-Required ${SYCL_BIN_DIR}/sycl?.dll $INSTALL_BIN_DIR
+Copy-Required ${SYCL_BIN_DIR}/ur_loader.dll $INSTALL_BIN_DIR
+Copy-Required ${SYCL_BIN_DIR}/ur_adapter_level_zero.dll $INSTALL_BIN_DIR
+Copy-Required ${SYCL_BIN_DIR}/ur_adapter_level_zero_v?.dll $INSTALL_BIN_DIR
+Copy-Required ${SYCL_BIN_DIR}/ur_win_proxy_loader.dll $INSTALL_BIN_DIR
 
 # sign
 ;& $env:SIGN_FILE_WINDOWS -q -vv (Get-ChildItem $INSTALL_BIN_DIR\* | Select-Object -Expand FullName)